+++ coda-src/portmapper/simpleclient.c
@@ -48,7 +48,7 @@ int main(void)
 	RPC2_PortIdent pident;
 	RPC2_SubsysIdent sident;
 	RPC2_CountedBS cident;
-	long	rc;
+	int	rc;
 	RPC2_Handle	cid;
 	RPC2_Integer	port;
 	struct timeval	timeout;
@@ -86,36 +86,36 @@ int main(void)
 
 	rc = portmapper_client_lookup_pbynvp(cid, "hithereservice", 0, 17, &port);
 
-	fprintf(stderr, "After initial lookup, rc=%ld, port=%ld\n", rc, port);
+	fprintf(stderr, "After initial lookup, rc=%d, port=%d\n", rc, port);
 
 	/* register the port */
 	rc = portmapper_client_register_excl(cid, "scraw", 0, 17, 12345);
 
-	fprintf(stderr, "After register, rc=%ld\n", rc);
+	fprintf(stderr, "After register, rc=%d\n", rc);
 
 	rc = portmapper_client_lookup_pbynvp(cid, "scraw", 0, 17, &port);
 
-	fprintf(stderr, "After second lookup, rc=%ld, port=%ld\n", rc, port);
+	fprintf(stderr, "After second lookup, rc=%d, port=%d\n", rc, port);
 
 	rc = portmapper_client_delete(cid, "scraw", 0, 17);
 
-	fprintf(stderr, "After delete, rc=%ld\n", rc);
+	fprintf(stderr, "After delete, rc=%d\n", rc);
 
 	rc = portmapper_client_register_excl(cid, "scraw", 0, 17, 12345);
 
-	fprintf(stderr, "After second register, rc=%ld\n", rc);
+	fprintf(stderr, "After second register, rc=%d\n", rc);
 
 	rc = portmapper_client_register_excl(cid, "scraw", 0, 17, 12345);
 
-	fprintf(stderr, "After third register, rc=%ld\n", rc);
+	fprintf(stderr, "After third register, rc=%d\n", rc);
 
 	rc = portmapper_client_register_sqsh(cid, "scraw", 0, 17, 12346);
 
-	fprintf(stderr, "After fourth register (sqsh), rc=%ld\n", rc);
+	fprintf(stderr, "After fourth register (sqsh), rc=%d\n", rc);
 
 	rc= portmapper_client_lookup_pbynvp(cid, "scraw", 0, 17, &port);
 
-	fprintf(stderr, "After third lookup, rc=%ld, port=%ld\n", rc, port);
+	fprintf(stderr, "After third lookup, rc=%d, port=%d\n", rc, port);
 	
 	return 0;
 }
